create custom hook 'usePreviousCount'
